+++ js/packages/store/messenger-effects/openingDaemonAndClients.effect.ts
@@ -1,33 +1,22 @@
 import { grpc } from '@improbable-eng/grpc-web'
-import { Dictionary } from '@reduxjs/toolkit'
 import { EventEmitter } from 'events'
-import i18next from 'i18next'
 import { Platform } from 'react-native'
-import RNFS from 'react-native-fs'
 
 import beapi from '@berty/api'
 import { GRPCError, createServiceClient } from '@berty/grpc-bridge'
 import { logger } from '@berty/grpc-bridge/middleware'
 import { bridge as rpcBridge, grpcweb as rpcWeb } from '@berty/grpc-bridge/rpc'
 import { deserializeFromBase64 } from '@berty/grpc-bridge/rpc/utils'
-import { ServiceClientType } from '@berty/grpc-bridge/welsh-clients.gen'
 import {
 	WelshMessengerServiceClient,
 	WelshProtocolServiceClient,
 } from '@berty/grpc-bridge/welsh-clients.gen'
-import { detectOSLanguage } from '@berty/i18n'
-import { GoBridge } from '@berty/native-modules/GoBridge'
 import { streamEventToAction as streamEventToReduxAction } from '@berty/redux/messengerActions'
-import { MessengerState } from '@berty/redux/reducers/messenger.reducer'
-import { resetTheme } from '@berty/redux/reducers/theme.reducer'
 import {
 	MESSENGER_APP_STATE,
 	setNextAccount,
 	setStateClients,
-	setStateOnBoardingReady,
 	setStateOpening,
-	setStateReady,
-	setStateSetupFinished,
 	setStateStreamDone,
 	setStateStreamInProgress,
 	setStreamError,
@@ -35,7 +24,6 @@ import {
 } from '@berty/redux/reducers/ui.reducer'
 import { AppDispatch, persistor } from '@berty/redux/store'
 import { accountClient, storageGet } from '@berty/utils/accounts/accountClient'
-import { updateAccount, refreshAccountList } from '@berty/utils/accounts/accountUtils'
 import { defaultCLIArgs } from '@berty/utils/accounts/defaultCLIArgs'
 import { convertMAddr } from '@berty/utils/ipfs/convertMAddr'
 import { requestAndPersistPushToken } from '@berty/utils/notification/notif-push'
@@ -86,80 +74,6 @@ const openAccountWithProgress = async (
 	}
 }
 
-const initBridge = async () => {
-	try {
-		console.log('bridge methods: ', Object.keys(GoBridge))
-		await GoBridge.initBridge()
-		console.log('bridge init done')
-	} catch (err: any) {
-		if (err?.message?.indexOf('already started') === -1) {
-			console.error('unable to init bridge: ', err)
-		} else {
-			console.log('bridge already started: ', err)
-		}
-	}
-}
-
-export const initialLaunch = async (dispatch: AppDispatch) => {
-	await initBridge()
-	const f = async () => {
-		const accounts = await refreshAccountList()
-
-		if (Object.keys(accounts).length > 0) {
-			let accountSelected: any = null
-			Object.values(accounts).forEach(account => {
-				if (!accountSelected) {
-					accountSelected = account
-				} else if (accountSelected && accountSelected.lastOpened < (account.lastOpened || 0)) {
-					accountSelected = account
-				}
-			})
-
-			// Delete berty-backup account
-			const outFile =
-				RNFS.TemporaryDirectoryPath + `/berty-backup-${accountSelected.accountId}` + '.tar'
-			RNFS.unlink(outFile)
-				.then(() => {
-					console.log('File deleted')
-				})
-				.catch(() => {
-					console.log('File berty backup does not exist') // here
-				})
-
-			dispatch(setNextAccount(accountSelected.accountId))
-
-			return
-		} else {
-			// this is the first account that will be created
-			dispatch(setStateOnBoardingReady())
-		}
-	}
-
-	f().catch(e => console.warn(e))
-}
-
-export const openingDaemonAndClients = async (
-	ui: UiState,
-	eventEmitter: EventEmitter,
-	dispatch: AppDispatch,
-) => {
-	if (ui.appState !== MESSENGER_APP_STATE.TO_OPEN) {
-		return
-	}
-
-	dispatch(setStateOpening())
-
-	try {
-		// opening berty daemon
-		await openingDaemon(ui.selectedAccount, dispatch)
-
-		// opening messenger and protocol clients
-		await openingClients(eventEmitter, dispatch)
-	} catch (err) {
-		console.warn(err)
-	}
-}
-
 const openingDaemon = async (selectedAccount: string | null, dispatch: AppDispatch) => {
 	if (selectedAccount === null) {
 		console.warn('openingDaemon / no account opened')
@@ -194,58 +108,6 @@ const openingDaemon = async (selectedAccount: string | null, dispatch: AppDispat
 	} catch (e) {
 		console.log(`account seems to be unopened yet ${e}`)
 	}
-}
-
-const openingClients = async (eventEmitter: EventEmitter, dispatch: AppDispatch): Promise<void> => {
-	console.log('starting stream')
-
-	let messengerClient, protocolClient
-
-	if (Platform.OS === 'web') {
-		const openedAccount = await accountClient?.getOpenedAccount({})
-		const url = convertMAddr(openedAccount?.listeners || [])
-
-		if (url === null) {
-			console.error('unable to find service address')
-			return
-		}
-
-		const opts = {
-			transport: grpc.CrossBrowserHttpTransport({ withCredentials: false }),
-			host: url,
-		}
-
-		protocolClient = createServiceClient(
-			beapi.protocol.ProtocolService,
-			rpcWeb(opts),
-		) as unknown as WelshProtocolServiceClient
-
-		messengerClient = createServiceClient(
-			beapi.messenger.MessengerService,
-			rpcWeb(opts),
-		) as unknown as WelshMessengerServiceClient
-	} else {
-		messengerClient = createServiceClient(
-			beapi.messenger.MessengerService,
-			rpcBridge,
-			logger.create('MESSENGER'),
-		)
-		protocolClient = createServiceClient(
-			beapi.protocol.ProtocolService,
-			rpcBridge,
-			logger.create('PROTOCOL'),
-		)
-	}
-
-	// request push notifications token
-	if (Platform.OS === 'ios' || Platform.OS === 'android') {
-		requestAndPersistPushToken(protocolClient).catch(e => console.warn(e))
-	}
-
-	// call messenger client event stream
-	messengerEventStream(messengerClient, eventEmitter, dispatch)
-
-	dispatch(setStateClients({ messengerClient, protocolClient }))
 }
 
 const messengerEventStream = (
@@ -332,72 +194,76 @@ const messengerEventStream = (
 		})
 }
 
-export const finishPreparingAccount = async (
-	ui: UiState,
-	messenger: MessengerState,
-	conversations: Dictionary<beapi.messenger.IConversation>,
-	dispatch: AppDispatch,
-) => {
-	try {
-		await closeConvos(ui.messengerClient, conversations)
+const openingClients = async (eventEmitter: EventEmitter, dispatch: AppDispatch): Promise<void> => {
+	console.log('starting stream')
 
-		await i18next.changeLanguage(detectOSLanguage())
+	let messengerClient, protocolClient
 
-		if (ui.isNewAccount) {
-			await updateAccountOnClients(ui.messengerClient, ui.selectedAccount, messenger.account)
-			// reset ui theme
-			dispatch(resetTheme())
-			dispatch(setStateSetupFinished())
-		} else {
-			dispatch(setStateReady())
+	if (Platform.OS === 'web') {
+		const openedAccount = await accountClient?.getOpenedAccount({})
+		const url = convertMAddr(openedAccount?.listeners || [])
+
+		if (url === null) {
+			console.error('unable to find service address')
+			return
 		}
-	} catch (err) {
-		console.warn(err)
+
+		const opts = {
+			transport: grpc.CrossBrowserHttpTransport({ withCredentials: false }),
+			host: url,
+		}
+
+		protocolClient = createServiceClient(
+			beapi.protocol.ProtocolService,
+			rpcWeb(opts),
+		) as unknown as WelshProtocolServiceClient
+
+		messengerClient = createServiceClient(
+			beapi.messenger.MessengerService,
+			rpcWeb(opts),
+		) as unknown as WelshMessengerServiceClient
+	} else {
+		messengerClient = createServiceClient(
+			beapi.messenger.MessengerService,
+			rpcBridge,
+			logger.create('MESSENGER'),
+		)
+		protocolClient = createServiceClient(
+			beapi.protocol.ProtocolService,
+			rpcBridge,
+			logger.create('PROTOCOL'),
+		)
 	}
+
+	// request push notifications token
+	if (Platform.OS === 'ios' || Platform.OS === 'android') {
+		requestAndPersistPushToken(protocolClient).catch(e => console.warn(e))
+	}
+
+	// call messenger client event stream
+	messengerEventStream(messengerClient, eventEmitter, dispatch)
+
+	dispatch(setStateClients({ messengerClient, protocolClient }))
 }
 
-const closeConvos = async (
-	client: ServiceClientType<beapi.messenger.MessengerService> | null,
-	conversations: { [key: string]: beapi.messenger.IConversation | undefined },
+export const openingDaemonAndClients = async (
+	ui: UiState,
+	eventEmitter: EventEmitter,
+	dispatch: AppDispatch,
 ) => {
-	if (client === null) {
-		console.warn('client is null')
+	if (ui.appState !== MESSENGER_APP_STATE.TO_OPEN) {
 		return
 	}
 
-	for (const conv of Object.values(conversations).filter(conv => conv?.isOpen) as any) {
-		client.conversationClose({ groupPk: conv.publicKey }).catch((e: any) => {
-			console.warn(`failed to close conversation "${conv.displayName}",`, e)
-		})
-	}
-}
+	dispatch(setStateOpening())
 
-const updateAccountOnClients = async (
-	messengerClient: ServiceClientType<beapi.messenger.MessengerService> | null,
-	selectedAccount: string | null,
-	account: beapi.messenger.IAccount | null | undefined,
-) => {
 	try {
-		// remove the displayName value that was set at the creation of the account
-		const displayName = await storageGet(GlobalPersistentOptionsKeys.DisplayName)
-		await storageRemove(GlobalPersistentOptionsKeys.DisplayName)
-		if (displayName) {
-			// update account in messenger client
-			await messengerClient?.accountUpdate({ displayName })
-			// update account in account client
-			await updateAccount({
-				accountName: displayName,
-				accountId: selectedAccount,
-				publicKey: account?.publicKey,
-			})
-		}
+		// opening berty daemon
+		await openingDaemon(ui.selectedAccount, dispatch)
+
+		// opening messenger and protocol clients
+		await openingClients(eventEmitter, dispatch)
 	} catch (err) {
 		console.warn(err)
 	}
-
-	// TODO: fix flow of asking permissions
-	// const config = await accountService.networkConfigGet({ accountId: selectedAccount })
-	// if (config.currentConfig?.staticRelay && config.currentConfig?.staticRelay[0] === ':default:') {
-	// 	await servicesAuthViaURL(protocolClient, bertyOperatedServer)
-	// }
 }
